requirements.txt
Flask==3.1.3
supabase==2.10.0
python-dotenv==1.0.1
gunicorn==21.2.0
